backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
email-validator
openai>=1.0.0
pillow
stripe>=7.0.0
